refactor: 이미지 repo interface 위치 변경
