+++ domain/src/main/java/com/shypolarbear/domain/usecase/image/ImageModifyUseCase.kt
@@ -2,7 +2,7 @@ package com.shypolarbear.domain.usecase.image
 
 import com.shypolarbear.domain.model.image.ImageModifyRequest
 import com.shypolarbear.domain.model.image.ImageModifyResponse
-import com.shypolarbear.domain.repository.ImageEditRepo
+import com.shypolarbear.domain.repository.image.ImageEditRepo
 
 class ImageModifyUseCase(
     private val repo: ImageEditRepo
